+++ e2e/e2etests/test_bitcoin_std_deposit_and_call_revert_other_address.go
@@ -38,7 +38,7 @@ func TestBitcoinStdMemoDepositAndCallRevertOtherAddress(r *runner.E2ERunner, arg
 
 	// ACT
 	// Deposit
-	txHash := r.DepositBTCWithAmount(amount, memo)
+	txHash := r.DepositBTCWithExactAmount(amount, memo)
 
 	// ASSERT
 	// Now we want to make sure revert TX is completed.